read shared

+++ rmosh/src/objects.rs
@@ -33,7 +33,7 @@ pub enum Object {
     Eof,
     EqHashtable(GcRef<EqHashtable>),
     False,
-    DefinedShared(isize),
+    DefinedShared(u32),
     FileInputPort(GcRef<FileInputPort>),
     FileOutputPort(GcRef<FileOutputPort>),
     Fixnum(isize),
